Use entries in Docker config, and account for dockerhub
The entries in the Docker config are base64 encoded credentials; I had
guessed they were outright tokens (i.e., something you can just stick
in a header as-is).

Docker Hub acts as an _index_ rather than a registry, so we have to do
an extra step to get a token for the actual registry.

Now you can do

    flumux list weaveworks/scope

and get an answer, since the registry defaults to Docker Hub and it
now does the extra token step.